Added Queries to UserRepository

+++ src/main/java/com/github/sirlacky/GiveBack/domain/repositories/RoleRepository.java
@@ -4,10 +4,11 @@ import com.github.sirlacky.GiveBack.domain.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface RoleRepository extends JpaRepository<Role,Integer> {
+public interface RoleRepository extends JpaRepository<Role,Long> {
 
     Role findByName(String name);
-
 
 }
